project: Move AT module to standalone project
