+++ project/at/src/cmds/at/at_test.c
@@ -20,9 +20,9 @@
 #include <termios.h>
 #include <unistd.h>
 
-#include <drivers/at/at_client.h>
-#include <drivers/at/at_emulator.h>
-#include <drivers/at/at_parser.h>
+#include <at/at_client.h>
+#include <at/emulator/at_emulator.h>
+#include <at/at_parser.h>
 
 /* PTY support */
 extern int ppty(int ptyfds[2]);
