+++ project/at/src/tests/at/at_parser_test.c
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <drivers/at/at_parser.h>
+#include <at/at_parser.h>
 #include <embox/test.h>
 
 EMBOX_TEST_SUITE("AT Parser Tests");
